some more fixes
