--- strassengezwitscher/strassengezwitscher/utils.py
@@ -1,5 +1,0 @@
-from rest_framework.views import exception_handler
-
-
-def custom_exception_handler(exc, context):
-    return exception_handler(exc, context)
